% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasty.R
\name{backend_hasty}
\alias{backend_hasty}
\title{Build a target using "hasty" parallelism}
\usage{
backend_hasty(config)
}
\description{
For internal use only
}
\examples{
\dontrun{
drake::test_with_dir("generates files", {
library(drake)
library(drake.hasty)
load_mtcars_example()
make(my_plan, parallelism = backend_hasty, jobs = 1)
make(
if (requireNamespace("clustermq")){
  options(clustermq.scheduler = "multicore")
  make(my_plan, parallelism = backend_hasty, jobs = 2)
}
})
}
}
\keyword{internal}
