% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasty.R
\name{default_hasty_build}
\alias{default_hasty_build}
\title{Build a target using "hasty" parallelism}
\usage{
default_hasty_build(target, config)
}
\arguments{
\item{target}{character, name of the target to build}

\item{config}{a \code{drake_config()} object}
}
\description{
For internal use only
}
