% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{drake.hasty}
\alias{drake.hasty}
\alias{drake.hasty-package}
\title{Hasty mode for the drake R package.}
\description{
Hasty mode is accelerated execution
with all of \code{drake}'s storage and reproducibility
guarantees stripped away. For experimentation only.
Use at your own risk.
}
\examples{
# See <https://github.com/wlandau/drake.hasty/blob/master/README.md>
# for examples.
}
\references{
\url{https://github.com/wlandau/drake.hasty}
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
