% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{hasty_build_default}
\alias{hasty_build_default}
\title{Build a target in hasty mode.}
\usage{
hasty_build_default(target, config)
}
\arguments{
\item{target}{character, name of the target to build}

\item{config}{a \code{drake_config()} object}
}
\description{
Runs really fast.
}
