% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{hasty_build_store}
\alias{hasty_build_store}
\title{Build a target in hasty mode with storage.}
\usage{
hasty_build_store(target, config)
}
\arguments{
\item{target}{character, name of the target to build}

\item{config}{a \code{drake_config()} object}
}
\description{
Builds the target, caches it, and does nothing else.
}
