% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{hasty_make}
\alias{hasty_make}
\title{Hasty mode for the drake R package}
\usage{
hasty_make(config)
}
\arguments{
\item{config}{a \code{drake_config()} object}
}
\description{
Sacrifices reproducibility to gain speed.
Targets are never cached and never up to date.
Use at your own risk.
}
\examples{
# See <https://github.com/wlandau/drake.hasty/blob/master/README.md>
# for examples.
}
