% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{remote_hasty_build}
\alias{remote_hasty_build}
\title{Build a target on a remote worker using "hasty" parallelism}
\usage{
remote_hasty_build(target, deps = NULL, config)
}
\arguments{
\item{target}{character, name of the target to build}

\item{deps}{named list of dependencies}

\item{config}{a \code{drake_config()} object}
}
\description{
For internal use only
}
\keyword{internal}
