% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasty.R
\name{remote_hasty_build}
\alias{remote_hasty_build}
\title{Build a target on a remote worker using "hasty" parallelism}
\usage{
remote_hasty_build(target, deps = NULL, config)
}
\arguments{
\item{deps}{named list of dependencies}
}
\description{
For internal use only
}
\keyword{internal}
